refactor(backend): consolidate files logic in module
